[libafl_qemu] EasyElf::resolve_symbol return GuestAddr
Also enforce Linux support at the crate level instead of item by item
